+++ package.json
@@ -1,10 +1,10 @@
 {
-  "name": "reactjs",
+  "name": "react-ts",
   "private": true,
   "version": "0.0.0",
   "scripts": {
     "dev": "vite",
-    "build": "vite build",
+    "build": "tsc && vite build",
     "preview": "vite preview"
   },
   "dependencies": {
@@ -12,7 +12,10 @@
     "react-dom": "^17.0.2"
   },
   "devDependencies": {
+    "@types/react": "^17.0.33",
+    "@types/react-dom": "^17.0.10",
     "@vitejs/plugin-react": "^1.0.7",
+    "typescript": "^4.5.4",
     "vite": "^2.8.0"
   }
 }
