--- src/components/age.jsx
@@ -1,2 +1,0 @@
-const ShowAge = props => <h2>Tuoi cua ban la: {props.age}</h2>
-export default ShowAge;
