+++ src/type/user.ts
@@ -1,0 +1,4 @@
+export type  userInfo = {
+    name: String,
+    age: Number
+}
